+++ routes/web.php
@@ -37,12 +37,12 @@ Route::group(['namespace' => 'App\Http\Controllers'], function()
             /**
              * User Routes
              */
-            Route::group(['prefix' => 'users'], function() {
+            Route::group(['middleware'=>'adminAccess','prefix' => 'users'], function() {
                 Route::get('/', 'UsersController@index')->name('users.index');
                 Route::get('/create', 'UsersController@create')->name('users.create');
                 Route::post('/create', 'UsersController@store')->name('users.store');
                 Route::get('/{user}/show', 'UsersController@show')->name('users.show');
-                Route::get('/{user}/observations', 'UsersController@observations')->name('users.observations');
+                Route::get('/{user}/observations', 'UsersController@observations')->name('users.observations');            
                 Route::get('/{user}/edit', 'UsersController@edit')->name('users.edit');
                 Route::patch('/{user}/update', 'UsersController@update')->name('users.update');
                 Route::delete('/{user}/delete', 'UsersController@destroy')->name('users.destroy');
@@ -53,12 +53,14 @@ Route::group(['namespace' => 'App\Http\Controllers'], function()
                 Route::get('/{user}/edit_user_courses', 'UsersController@edit_user_courses')->name('users.edit_user_courses');
                 Route::patch('/{user}/update_user_courses', 'UsersController@update_user_courses')->name('users.update_user_courses');
                 Route::delete('/{user}/courses_teach/{course}/destroy_user_courses', 'UsersController@destroy_user_courses')->name('users.destroy_user_courses');
+
+                Route::patch('/setObservations', 'UsersController@setObservations')->name('users.setObservations');
             });
 
             /**
              * Course Routes
              */
-            Route::group(['prefix' => 'courses'], function() {
+            Route::group(['middleware'=>'adminAccess','prefix' => 'courses'], function() {
                 Route::get('/', 'CoursesController@index')->name('courses.index');
                 Route::get('/create', 'CoursesController@create')->name('courses.create');
                 Route::post('/create', 'CoursesController@store')->name('courses.store');
@@ -73,7 +75,7 @@ Route::group(['namespace' => 'App\Http\Controllers'], function()
             /**
              * Room Routes
              */
-            Route::group(['prefix' => 'rooms'], function() {
+            Route::group(['middleware'=>'adminAccess','prefix' => 'rooms'], function() {
                 Route::get('/index', 'RoomsController@index')->name('rooms.index');
                 Route::get('/create', 'RoomsController@create')->name('rooms.create');
                 Route::post('/create', 'RoomsController@store')->name('rooms.store');
@@ -89,18 +91,23 @@ Route::group(['namespace' => 'App\Http\Controllers'], function()
              */
             Route::group(['prefix' => 'rotations'], function() {
                 //Rotation CURD
-                Route::get('/index', 'RotationsController@index')->name('rotations.index');
-                Route::get('/create', 'RotationsController@create')->name('rotations.create');
-                Route::post('/create', 'RotationsController@store')->name('rotations.store');
-                Route::get('/{rotation}/edit', 'RotationsController@edit')->name('rotations.edit');
-                Route::patch('/{rotation}/update', 'RotationsController@update')->name('rotations.update');
-                Route::delete('/{rotation}/delete', 'RotationsController@destroy')->name('rotations.destroy');
+                Route::get('/index', 'RotationsController@index')->name('rotations.index')->middleware('adminAccess');
+                Route::get('/create', 'RotationsController@create')->name('rotations.create')->middleware('adminAccess');
+                Route::post('/create', 'RotationsController@store')->name('rotations.store')->middleware('adminAccess');
+                Route::get('/{rotation}/edit', 'RotationsController@edit')->name('rotations.edit')->middleware('adminAccess');
+                Route::patch('/{rotation}/update', 'RotationsController@update')->name('rotations.update')->middleware('adminAccess');
+                Route::delete('/{rotation}/delete', 'RotationsController@destroy')->name('rotations.destroy')->middleware('adminAccess');
                 //Rotation CURD
                 //Rotation ExamProgram
                 Route::get('/{rotation}/show', 'CourseRotation_ExamProgramController@show')->name('rotations.program.show');
-                Route::get('/{rotation}/create', 'CourseRotation_ExamProgramController@add_course_to_the_program')->name('rotations.program.add_course_to_the_program');
-                Route::post('/{rotation}/store', 'CourseRotation_ExamProgramController@store_course_to_the_program')->name('rotations.program.store_course_to_the_program');
-                Route::get('/{rotation}/course/{course}/delete_course_from_program', 'CourseRotation_ExamProgramController@delete_course_from_program')->name('rotations.course.delete_course_from_program');
+                Route::get('/{rotation}/create', 'CourseRotation_ExamProgramController@add_course_to_the_program')->name('rotations.program.add_course_to_the_program')->middleware('adminAccess');
+                Route::post('/{rotation}/store', 'CourseRotation_ExamProgramController@store_course_to_the_program')->name('rotations.program.store_course_to_the_program')->middleware('adminAccess');
+                Route::delete('/{rotation}/course/{course}/delete_course_from_program', 'CourseRotation_ExamProgramController@delete_course_from_program')->name('rotations.course.delete_course_from_program')->middleware('adminAccess');
+                //initialization --> 1. initExamProgram, 2. initRoomsInAllCourses , 3. initUsersObservationsInAllCourses
+                Route::get('/{rotation}/initExamProgram', 'CourseRotation_ExamProgramController@initExamProgram')->name('rotations.initExamProgram')->middleware('adminAccess');
+                Route::get('/{rotation}/initRoomsInAllCourses', 'CourseRotation_ExamProgramController@initRoomsInAllCourses')->name('rotations.initRoomsInAllCourses')->middleware('adminAccess');
+                Route::get('/{rotation}/initUsersObservationsInAllCourses', 'CourseRotation_ExamProgramController@initUsersObservationsInAllCourses')->name('rotations.initUsersObservationsInAllCourses')->middleware('adminAccess');
+
                 //Rotation ExamProgram
                 //Rotation Objections
                 Route::get('/{rotation}/objections/create', 'CourseRotationUser_ObjectionController@create')->name('rotations.objections.create');
@@ -109,20 +116,22 @@ Route::group(['namespace' => 'App\Http\Controllers'], function()
                 Route::patch('/{rotation}/objections/update', 'CourseRotationUser_ObjectionController@update')->name('rotations.objections.update');
                 //Rotation Objections
                 //Distribute Students && Members of Faculty
-                Route::post('/{rotation}/distributeStudents', 'RotationsController@distributeStudents')->name('rotations.distributeStudents');
-                Route::post('/{rotation}/distributeMembersOfFaculty', 'RotationsController@distributeMembersOfFaculty')->name('rotations.distributeMembersOfFaculty');
+                Route::post('/{rotation}/distributeStudents', 'RotationsController@distributeStudents')->name('rotations.distributeStudents')->middleware('adminAccess');
+                Route::post('/{rotation}/distributeMembersOfFaculty', 'RotationsController@distributeMembersOfFaculty')->name('rotations.distributeMembersOfFaculty')->middleware('adminAccess');
 
                 //initial_members
-                Route::get('/{rotation}/initial_members', 'RotationsController@create_initial_members')->name('rotations.create_initial_members');
-                Route::post('/{rotation}/initial_members', 'RotationsController@store_initial_members')->name('rotations.store_initial_members');
-                Route::get('/{rotation}/edit_initial_members', 'RotationsController@edit_initial_members')->name('rotations.edit_initial_members');
-                Route::patch('/{rotation}/update_initial_members', 'RotationsController@update_initial_members')->name('rotations.update_initial_members');
+                Route::get('/{rotation}/initial_members', 'RotationsController@create_initial_members')->name('rotations.create_initial_members')->middleware('adminAccess');
+                Route::post('/{rotation}/initial_members', 'RotationsController@store_initial_members')->name('rotations.store_initial_members')->middleware('adminAccess');
+                Route::get('/{rotation}/edit_initial_members', 'RotationsController@edit_initial_members')->name('rotations.edit_initial_members')->middleware('adminAccess');
+                Route::patch('/{rotation}/update_initial_members', 'RotationsController@update_initial_members')->name('rotations.update_initial_members')->middleware('adminAccess');
+
+            
             });
                         
             /**
              * Course with Rotation Routes
              */
-            Route::group(['prefix' => 'rotations'], function() {
+            Route::group(['middleware'=>'adminAccess','prefix' => 'rotations'], function() {
                 Route::get('/{rotation}/course/{course}/room/{specific_room}', 'CourseRotationController@get_room_for_course')->name('rotations.get_room_for_course');
                 Route::patch('/{rotation}/course/{course}/room/{specific_room}', 'CourseRotationController@customize_room_for_course')->name('rotations.customize_room_for_course');
                 Route::get('/{rotation}/course/{course}/show', 'CourseRotationController@show')->name('rotations.course.show');
@@ -133,9 +142,13 @@ Route::group(['namespace' => 'App\Http\Controllers'], function()
              * Course with Objection Routes
              */
             Route::group(['prefix' => 'objections'], function() {
-                Route::get('/objections/user/{user}/index', 'CourseRotationUser_ObjectionController@index')->name('objections.user.index');
+                Route::get('/user/{user}/index', 'CourseRotationUser_ObjectionController@index')->name('objections.user.index')->middleware('adminAccess');
+                Route::get('/rotation/{rotation}/user/{user}/show', 'CourseRotationUser_ObjectionController@show')->name('objections.user.show')->middleware('adminAccess');
             });
-
+            Route::group(['prefix' => 'observations'], function() {
+                //to show observation for user in spesific rotation
+                Route::get('rotations/{rotation}/user/{user}/show', 'RotationsController@showObservationsInSpecificRotationForSpecificUser')->name('rotations.observations.user.show')->middleware('adminAccess');
+            });
         });
     });
